requirements.txt
google-cloud-pubsub
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2